[refactor/#43] 폴더 구조 정리
