+++ src/components/ReservationHistory/index.tsx
@@ -1,0 +1,1 @@
+export function ReservationHistory() {}
